src/main/java/ejercicios/AnnosBisiestos.java

package ejercicios;

import static ejercicios.SetDeTenis.evaluar;
import java.util.Scanner;

/**
 *
 * @author danielsanchez
 */
public class AnnosBisiestos {
    public static String evaluar(int anno){
        int abis = anno % 4;
        if(abis == 0){
            String respuesta = anno + " es bisiesto";
            return respuesta;
        }else{
            String respuesta = anno + " no es bisiesto";
            return respuesta;
        }
    }
    
    public static void main(String[] args) {
        Scanner lector = new Scanner(System.in);
        System.out.print("Año:");
        int anno = lector.nextInt();
        
        String respuesta = evaluar(anno);
        System.out.println(respuesta);
    }
            
}
